refactor: blog/slide テーブルの公開トグルを PublishToggle に共通化
- 楽観的更新（失敗時に状態を戻す）を含む公開/下書きトグルのロジックを
  app/(authenticated)/_components の PublishToggle に集約
- BlogRow / SlideRow は行レイアウトのみを持つ薄い表示に
