+++ apps/admin/src/app/(authenticated)/_components/publish-toggle/index.ts
@@ -1,0 +1,1 @@
+export { PublishToggle } from './publish-toggle';
